Retire mtime-based directory gating from the library scan
The scan skipped directories whose mtime predated the folder's
last_checked_at. That cannot survive alongside removal: renaming a
directory updates its parent's mtime and never its own, so a renamed
subtree keeps an old timestamp, last_checked_at only moves forward, and a
gated walk never looks at it again. Every image under a renamed folder
would be treated as deleted and never found.

The gate bought little anyway. os.walk has already listed each directory
by the time the check runs, so it added a getmtime() per directory and
saved only a filename suffix check; the costly work is already avoided by
diffing against the known catalog paths.

last_checked_at stays, since the Library page shows it and it is the only
evidence a sync ran when nothing changed. Only its gating role goes, which
also makes clearing it on a path update pointless.
